Changed numerous things within the App
