--- .Rprofile
@@ -1,1 +1,0 @@
-source("renv/activate.R")
