+++ FrequencyBandAnalysis/.Rprofile
@@ -1,0 +1,1 @@
+source("renv/activate.R")
