+++ FrequencyBandAnalysis/Sample Data/Multivariate Data/desktop.ini
@@ -1,0 +1,6 @@
+[LocalizedFileNames]
+MV_S3B_Data.csv=@MV_S3B_Data,0
+MV_L3B_Data.csv=@MV_L3B_Data,0
+MV_M3B1_Data.csv=@MV_M3B1_Data,0
+MV_WN_Data.csv=@MV_WN_Data,0
+MV_M3B2_Data.csv=@MV_M3B2_Data,0
