--- README.md
@@ -1,3 +1,0 @@
-# Summer-Visualization-Project
-
-Add some details here!
